Update zigo to 0.12.0, spell codepoints as rune, add colors, modes and selections
zigo 0.12.0 brings codepoint semantics: `.codepoints = .infer_u21` makes
every u21 parameter and return a Go rune, and the u32 codepoint sites
(GraphemeWidth, PrintSlice, RenderCell.Codepoint,
KeyEvent.UnshiftedCodepoint) opt in explicitly. Codepoint arguments are
now range-checked against Unicode, so GraphemeWidth returns an error.
Generated helpers gained a zigo prefix; the one test that reached for one
follows.

New API from the go-libghostty comparison:

- Colors as 0xRRGGBB, the packing RenderCell already uses: the dynamic
  background, foreground and cursor colors, the 256-color palette, and
  setters for the configured defaults.
- Mode as a text-parsable enum with ModeEnabled and SetMode. SetMode only
  flips state; the parser's side effects for screen switching and
  132-column stay behind SwitchScreenMode and Deccolm.
- Selection as a value in screen coordinates, read and set on a Screen
  and returned by Search.Matches and Search.SelectedMatch, with
  Screen.ViewportTop for converting to viewport rows.

The example moves to examples/hypercat inside the root module, so
`go build ./...` compiles it against the checkout and CI no longer needs
a replace directive. The root module's Go version rises to 1.25 because
the example's dependencies require it.

+++ examples/hypercat/emoji_test.go
@@ -93,7 +93,7 @@ func TestOnlyWideCellsAreEmoji(t *testing.T) {
 		wide[r] = false
 	}
 	for r, want := range wide {
-		w, err := gostty.CodepointWidth(uint32(r))
+		w, err := gostty.CodepointWidth(r)
 		if err != nil {
 			t.Fatalf("CodepointWidth(%q): %v", r, err)
 		}
